Internal PartitionKeyRangeCache: wire up force refresh on GetRoutingMapAsync
Batch in Friends needs to be able to ask for a refresh of the cached RID in the case where it receives a null (<=> not found) response from GetRoutingMapAsync, as it could mean the collection was recycled/deleted externally.

I am also switching a couple of methods written using ContinueWith to be async instead as I found during testing that this style can cause AggregateExceptions to be thrown which callers do not typically expect (e.g. a NotFound would end up as a 500 because of that).

Signed-off-by: Floran Narenji <flnarenj@microsoft.com>